tq-97: TableEditFilters
